feat: move OTel utility functions into `pkg/trace/transform` to break `pkg/trace` → `pkg/trace/otel` dependency (#47758)
## Problem

PR #44973 did a first step to isolate OTel collector dependencies from `dd-trace-go` by creating separate submodules, but there were some leftover import paths that kept `pkg/trace` depending on `pkg/trace/otel`. This caused `dd-trace-go` to stil transitively pull in heavy OTel collector packages.

## Changes

- Moves OTel utility functions from `pkg/trace/otel/traceutil` → `pkg/trace/transform/otelutil.go`
- Converts `pkg/trace/otel/traceutil` into a thin re-export shim for backward compatibility
- Moves OTel stats code from `pkg/trace/stats/otel.go` → `pkg/trace/otel/stats/otel_util.go`
- Updates `pkg/trace/api` and pkg/trace/transform to use local references
- Removes `pkg/trace/otel` from the require sections of `pkg/trace` and `pkg/trace/stats` go.mod

### Describe how you validated your changes

CI tests.
