+++ pkg/trace/api/otlp.go
@@ -30,7 +30,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/log"
-	traceutilotel "github.com/DataDog/datadog-agent/pkg/trace/otel/traceutil"
 	"github.com/DataDog/datadog-agent/pkg/trace/sampler"
 	"github.com/DataDog/datadog-agent/pkg/trace/semantics"
 	"github.com/DataDog/datadog-agent/pkg/trace/timing"
@@ -278,7 +277,7 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 
 	// Get the hostname or set to empty if source is empty
 	var hostname string
-	src, srcok := o.conf.OTLPReceiver.AttributesTranslator.ResourceToSource(ctx, rspans.Resource(), traceutilotel.SignalTypeSet, hostFromAttributesHandler)
+	src, srcok := o.conf.OTLPReceiver.AttributesTranslator.ResourceToSource(ctx, rspans.Resource(), transform.SignalTypeSet, hostFromAttributesHandler)
 	if srcok {
 		switch src.Kind {
 		case source.HostnameKind:
@@ -298,13 +297,13 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 	resAccessor := semantics.NewPDataMapAccessor(otelres.Attributes())
 
 	// Get container ID from OTel semantic conventions
-	containerID := traceutilotel.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptContainerID, true)
+	containerID := transform.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptContainerID, true)
 	if containerID == "" && !o.conf.HasFeature("enable_otlp_container_tags_v2") {
-		containerID = traceutilotel.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptK8sPodUID, true)
+		containerID = transform.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptK8sPodUID, true)
 	}
 
 	// Get env from OTel semantic conventions
-	env := traceutilotel.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptDeploymentEnv, true)
+	env := transform.LookupSemanticStringWithAccessor(resAccessor, semantics.ConceptDeploymentEnv, true)
 
 	// Get container tags from OTel semantic conventions
 	var containerTags string
@@ -341,7 +340,7 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 		for j := 0; j < libspans.Spans().Len(); j++ {
 			otelspan := libspans.Spans().At(j)
 			spancount++
-			traceID := traceutilotel.OTelTraceIDToUint64(otelspan.TraceID())
+			traceID := transform.OTelTraceIDToUint64(otelspan.TraceID())
 			if tracesByID[traceID] == nil {
 				tracesByID[traceID] = pb.Trace{}
 			}
@@ -354,11 +353,11 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 		}
 	}
 
-	lang := traceutilotel.GetOTelAttrVal(otelres.Attributes(), true, string(semconv.TelemetrySDKLanguageKey))
+	lang := transform.GetOTelAttrVal(otelres.Attributes(), true, string(semconv.TelemetrySDKLanguageKey))
 	tagstats := &info.TagStats{
 		Tags: info.Tags{
 			Lang:            lang,
-			TracerVersion:   "otlp-" + traceutilotel.GetOTelAttrVal(otelres.Attributes(), true, string(semconv.TelemetrySDKVersionKey)),
+			TracerVersion:   "otlp-" + transform.GetOTelAttrVal(otelres.Attributes(), true, string(semconv.TelemetrySDKVersionKey)),
 			EndpointVersion: "opentelemetry_grpc_v1",
 		},
 		Stats: info.NewStats(),
@@ -369,7 +368,7 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 	_ = o.statsd.Count("datadog.trace_agent.otlp.traces", int64(len(tracesByID)), tags, 1)
 	p := Payload{
 		Source:                 tagstats,
-		ClientComputedStats:    traceutilotel.GetOTelAttrVal(otelres.Attributes(), true, keyStatsComputed) != "" || clientComputedStats,
+		ClientComputedStats:    transform.GetOTelAttrVal(otelres.Attributes(), true, keyStatsComputed) != "" || clientComputedStats,
 		ClientComputedTopLevel: o.conf.HasFeature("enable_otlp_compute_top_level_by_span_kind"),
 		TracerPayload: &pb.TracerPayload{
 			Hostname:      hostname,
@@ -394,7 +393,7 @@ func (o *OTLPReceiver) receiveResourceSpansV2(ctx context.Context, rspans ptrace
 func (o *OTLPReceiver) receiveResourceSpansV1(ctx context.Context, rspans ptrace.ResourceSpans, httpHeader http.Header, hostFromAttributesHandler attributes.HostFromAttributesHandler) source.Source {
 	// each rspans is coming from a different resource and should be considered
 	// a separate payload; typically there is only one item in this slice
-	src, srcok := o.conf.OTLPReceiver.AttributesTranslator.ResourceToSource(ctx, rspans.Resource(), traceutilotel.SignalTypeSet, hostFromAttributesHandler)
+	src, srcok := o.conf.OTLPReceiver.AttributesTranslator.ResourceToSource(ctx, rspans.Resource(), transform.SignalTypeSet, hostFromAttributesHandler)
 	hostFromMap := func(m map[string]string, key string) {
 		// hostFromMap sets the hostname to m[key] if it is set.
 		if v, ok := m[key]; ok {
@@ -441,7 +440,7 @@ func (o *OTLPReceiver) receiveResourceSpansV1(ctx context.Context, rspans ptrace
 		for j := 0; j < libspans.Spans().Len(); j++ {
 			spancount++
 			span := libspans.Spans().At(j)
-			traceID := traceutilotel.OTelTraceIDToUint64(span.TraceID())
+			traceID := transform.OTelTraceIDToUint64(span.TraceID())
 			if tracesByID[traceID] == nil {
 				tracesByID[traceID] = pb.Trace{}
 			}
@@ -593,9 +592,9 @@ func (o *OTLPReceiver) createChunks(tracesByID map[uint64]pb.Trace, prioritiesBy
 func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.InstrumentationScope, in ptrace.Span) *pb.Span {
 	traceID := [16]byte(in.TraceID())
 	span := &pb.Span{
-		TraceID:  traceutilotel.OTelTraceIDToUint64(traceID),
-		SpanID:   traceutilotel.OTelSpanIDToUint64(in.SpanID()),
-		ParentID: traceutilotel.OTelSpanIDToUint64(in.ParentSpanID()),
+		TraceID:  transform.OTelTraceIDToUint64(traceID),
+		SpanID:   transform.OTelSpanIDToUint64(in.SpanID()),
+		ParentID: transform.OTelSpanIDToUint64(in.ParentSpanID()),
 		Start:    int64(in.StartTimestamp()),
 		Duration: int64(in.EndTimestamp()) - int64(in.StartTimestamp()),
 		Meta:     make(map[string]string, res.Attributes().Len()),
@@ -612,7 +611,7 @@ func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.Instrumenta
 	}
 
 	transform.SetMetaOTLP(span, "otel.trace_id", hex.EncodeToString(traceID[:]))
-	transform.SetMetaOTLP(span, "span.kind", traceutilotel.OTelSpanKindName(spanKind))
+	transform.SetMetaOTLP(span, "span.kind", transform.OTelSpanKindName(spanKind))
 	if _, ok := span.Meta["version"]; !ok {
 		if ver, ok := res.Attributes().Get(string(semconv.ServiceVersionKey)); ok && ver.AsString() != "" {
 			transform.SetMetaOTLP(span, "version", ver.AsString())
@@ -677,7 +676,7 @@ func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.Instrumenta
 
 	// Check for db.namespace and conditionally set db.name
 	if _, ok := span.Meta["db.name"]; !ok {
-		if dbNamespace := traceutilotel.GetOTelAttrValInResAndSpanAttrs(in, res, false, string(semconv127.DBNamespaceKey)); dbNamespace != "" {
+		if dbNamespace := transform.GetOTelAttrValInResAndSpanAttrs(in, res, false, string(semconv127.DBNamespaceKey)); dbNamespace != "" {
 			transform.SetMetaOTLP(span, "db.name", dbNamespace)
 		}
 	}
@@ -697,12 +696,12 @@ func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.Instrumenta
 	}
 	span.Error = transform.Status2Error(in.Status(), in.Events(), span.Meta)
 	if transform.OperationAndResourceNameV2Enabled(o.conf) {
-		span.Name = traceutilotel.GetOTelOperationNameV2(in, res)
+		span.Name = transform.GetOTelOperationNameV2(in, res)
 	} else {
 		if span.Name == "" {
 			name := in.Name()
 			if !o.conf.OTLPReceiver.SpanNameAsResourceName {
-				name = traceutilotel.OTelSpanKindName(in.Kind())
+				name = transform.OTelSpanKindName(in.Kind())
 				if lib.Name() != "" {
 					name = lib.Name() + "." + name
 				} else {
@@ -720,7 +719,7 @@ func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.Instrumenta
 	}
 	if span.Resource == "" {
 		if transform.OperationAndResourceNameV2Enabled(o.conf) {
-			span.Resource = traceutilotel.GetOTelResourceV2(in, res)
+			span.Resource = transform.GetOTelResourceV2(in, res)
 		} else {
 			if r := resourceFromTags(span.Meta); r != "" {
 				span.Resource = r
@@ -730,7 +729,7 @@ func (o *OTLPReceiver) convertSpan(res pcommon.Resource, lib pcommon.Instrumenta
 		}
 	}
 	if span.Type == "" {
-		span.Type = traceutilotel.SpanKind2Type(in, res)
+		span.Type = transform.SpanKind2Type(in, res)
 	}
 	return span
 }
